backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3

# ML
transformers==4.45.2
torch==2.4.1
sentence-transformers==3.2.0
chromadb==0.5.18

# Email / Google
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# HTML parsing
beautifulsoup4==4.12.3

# HTTP client
httpx==0.27.2
